+++ src/main/java/com/eodegano/cocobackend/service/PoiCurationService.java
@@ -3,5 +3,5 @@ package com.eodegano.cocobackend.service;
 import com.eodegano.cocobackend.dto.PoiCurationResponseDto;
 
 public interface PoiCurationService {
-    PoiCurationResponseDto getPoiList(String sigunguCode, Integer peopleCount, Integer contentTypeId);
+    PoiCurationResponseDto getPoiList(String sigunguCode, Integer peopleCount, Integer contentTypeId, String userEmail);
 }