+++ src/main/java/com/eodegano/cocobackend/service/PoiDetailService.java
@@ -3,5 +3,5 @@ package com.eodegano.cocobackend.service;
 import com.eodegano.cocobackend.dto.PoiDetailResponseDto;
 
 public interface PoiDetailService {
-    PoiDetailResponseDto getPoiDetail(Long contentId);
+    PoiDetailResponseDto getPoiDetail(Long contentId, String userEmail);
 }